+++ apps/ceos/src/pages/recruit/apply/index.tsx
@@ -1,13 +1,4 @@
-import {
-  Flex,
-  Text,
-  Button,
-  Desktop,
-  TextField,
-  SelectButton,
-  theme,
-  CheckBox,
-} from '@ceos-fe/ui';
+import { Flex, Text, Button, Desktop, theme } from '@ceos-fe/ui';
 import Information from './Information';
 import styled from '@emotion/styled';
 import {
@@ -29,6 +20,9 @@ import {
   useForm,
 } from 'react-hook-form';
 import { Title } from '../../../components/Title';
+import Common from './Common';
+import Part from './Part';
+import Schedule from './Schedule';
 
 export interface QuestionProps {
   questionId: number;
@@ -51,6 +45,7 @@ export interface RecruitApplyFormInterface {
   watch: UseFormWatch<RecruitApplyValuesInterface>;
   setValue: UseFormSetValue<RecruitApplyValuesInterface>;
   handleSubmit: UseFormHandleSubmit<RecruitApplyValuesInterface>;
+  questionList?: RecruitApplyResponse;
 }
 
 const Apply = () => {
@@ -87,8 +82,6 @@ const Apply = () => {
     console.log(data.part); // 현재 선택된 "part" 값 출력
   };
 
-  const selectedPart = watch('part');
-
   const questionList = data?.pages[0].data;
 
   console.log(questionList);
@@ -112,286 +105,30 @@ const Apply = () => {
             handleSubmit={handleSubmit}
           />
           {/* 공통 질문 */}
-          <Section>
-            <Text webTypo="Heading3" paletteColor="Blue">
-              공통 질문
-            </Text>
-            {questionList?.commonQuestions.map((ques, idx) => (
-              <Flex
-                direction="column"
-                align="start"
-                webGap={12}
-                key={ques.questionId}
-              >
-                <Text webTypo="Label3">{`${idx + 1}. ${ques.question}`}</Text>
-                <TextField width={856} multiline={ques.multiline} />
-                <Flex direction="column" align="start">
-                  {ques.questionDetail.map((detail, idx) =>
-                    detail.color === 'gray' ? (
-                      <Text
-                        webTypo="Body3"
-                        paletteColor="Gray5"
-                        key={`detail_${idx}`}
-                      >
-                        {detail.explaination}
-                      </Text>
-                    ) : (
-                      <Text
-                        webTypo="Body3"
-                        paletteColor="Blue"
-                        key={`detail_${idx}`}
-                      >
-                        {detail.explaination}
-                      </Text>
-                    ),
-                  )}
-                </Flex>
-              </Flex>
-            ))}
-            <RowLine />
-          </Section>
+          <Common
+            register={register}
+            watch={watch}
+            setValue={setValue}
+            handleSubmit={handleSubmit}
+            questionList={questionList}
+          />
           {/* 파트별 질문 */}
-          <Section>
-            <Flex direction="column" align="start" webGap={12}>
-              <Text webTypo="Heading3" paletteColor="Blue">
-                파트별 질문
-              </Text>
-              <Text webTypo="Body3" paletteColor="Gray5">
-                *지원하고자 하는 파트를 선택하여 답변을 작성해주세요.
-              </Text>
-            </Flex>
-            <form onSubmit={handleSubmit(onSubmit)}>
-              <Flex webGap={20}>
-                <SelectButton
-                  variant="ceos"
-                  value="기획"
-                  webWidth={205}
-                  {...register('part')}
-                />
-                <SelectButton
-                  variant="ceos"
-                  value="디자인"
-                  webWidth={205}
-                  {...register('part')}
-                />
-                <SelectButton
-                  variant="ceos"
-                  value="프론트엔드"
-                  webWidth={205}
-                  {...register('part')}
-                />
-                <SelectButton
-                  variant="ceos"
-                  value="백엔드"
-                  webWidth={205}
-                  {...register('part')}
-                />
-              </Flex>
-            </form>
-
-            <Section>
-              {selectedPart === '기획' ? (
-                <>
-                  {questionList?.productQuestions.map((ques, idx) => (
-                    <Flex
-                      direction="column"
-                      align="start"
-                      webGap={12}
-                      key={ques.questionId}
-                    >
-                      <Text webTypo="Label3">{`${idx + 1}. ${
-                        ques.question
-                      }`}</Text>
-                      <TextField width={856} multiline={ques.multiline} />
-                      <Flex direction="column" align="start">
-                        {ques.questionDetail.map((detail, idx) =>
-                          detail.color === 'gray' ? (
-                            <Text
-                              webTypo="Body3"
-                              paletteColor="Gray5"
-                              key={`detail_${idx}`}
-                            >
-                              {detail.explaination}
-                            </Text>
-                          ) : (
-                            <Text
-                              webTypo="Body3"
-                              paletteColor="Blue"
-                              key={`detail_${idx}`}
-                            >
-                              {detail.explaination}
-                            </Text>
-                          ),
-                        )}
-                      </Flex>
-                    </Flex>
-                  ))}
-                </>
-              ) : selectedPart === '디자인' ? (
-                <>
-                  {questionList?.designQuestions.map((ques, idx) => (
-                    <Flex
-                      direction="column"
-                      align="start"
-                      webGap={12}
-                      key={ques.questionId}
-                    >
-                      <Text webTypo="Label3">{`${idx + 1}. ${
-                        ques.question
-                      }`}</Text>
-                      <TextField width={856} multiline={ques.multiline} />
-                      <Flex direction="column" align="start">
-                        {ques.questionDetail.map((detail, idx) =>
-                          detail.color === 'gray' ? (
-                            <Text
-                              webTypo="Body3"
-                              paletteColor="Gray5"
-                              key={`detail_${idx}`}
-                            >
-                              {detail.explaination}
-                            </Text>
-                          ) : (
-                            <Text
-                              webTypo="Body3"
-                              paletteColor="Blue"
-                              key={`detail_${idx}`}
-                            >
-                              {detail.explaination}
-                            </Text>
-                          ),
-                        )}
-                      </Flex>
-                    </Flex>
-                  ))}
-                </>
-              ) : selectedPart === '프론트엔드' ? (
-                <>
-                  {questionList?.frontendQuestions.map((ques, idx) => (
-                    <Flex
-                      direction="column"
-                      align="start"
-                      webGap={12}
-                      key={ques.questionId}
-                    >
-                      <Text webTypo="Label3">{`${idx + 1}. ${
-                        ques.question
-                      }`}</Text>
-                      <TextField width={856} multiline={ques.multiline} />
-                      <Flex direction="column" align="start">
-                        {ques.questionDetail.map((detail, idx) =>
-                          detail.color === 'gray' ? (
-                            <Text
-                              webTypo="Body3"
-                              paletteColor="Gray5"
-                              key={`detail_${idx}`}
-                            >
-                              {detail.explaination}
-                            </Text>
-                          ) : (
-                            <Text
-                              webTypo="Body3"
-                              paletteColor="Blue"
-                              key={`detail_${idx}`}
-                            >
-                              {detail.explaination}
-                            </Text>
-                          ),
-                        )}
-                      </Flex>
-                    </Flex>
-                  ))}
-                </>
-              ) : selectedPart === '백엔드' ? (
-                <>
-                  {questionList?.backendQuestions.map((ques, idx) => (
-                    <Flex
-                      direction="column"
-                      align="start"
-                      webGap={12}
-                      key={ques.questionId}
-                    >
-                      <Text webTypo="Label3">{`${idx + 1}. ${
-                        ques.question
-                      }`}</Text>
-                      <TextField width={856} multiline={ques.multiline} />
-                      <Flex direction="column" align="start">
-                        {ques.questionDetail.map((detail, idx) =>
-                          detail.color === 'gray' ? (
-                            <Text
-                              webTypo="Body3"
-                              paletteColor="Gray5"
-                              key={`detail_${idx}`}
-                            >
-                              {detail.explaination}
-                            </Text>
-                          ) : (
-                            <Text
-                              webTypo="Body3"
-                              paletteColor="Blue"
-                              key={`detail_${idx}`}
-                            >
-                              {detail.explaination}
-                            </Text>
-                          ),
-                        )}
-                      </Flex>
-                    </Flex>
-                  ))}
-                </>
-              ) : (
-                <></>
-              )}
-            </Section>
-            <RowLine />
-          </Section>
-          <Flex
-            direction="column"
-            align="start"
-            width={856}
-            webGap={36}
-            margin={'24px 0 100px 0'}
-          >
-            <div>
-              <Text webTypo="Heading3" paletteColor="Blue">
-                면접 날짜
-              </Text>
-
-              <Text webTypo="Body3" paletteColor="Gray5" margin="8px 0 0 0">
-                *불가능한 날짜와 시간에 체크해주세요. 가능한 날짜가 아니라,
-                불가능한 날짜입니다.
-              </Text>
-              <Text webTypo="Body3" paletteColor="Gray5">
-                *모든 면접은 화상(ZOOM)으로 이루어지며, 면접 시작 10분 전에
-                대기실 참가 안내를 드리니 이를 고려하여 선택 부탁드립니다.
-              </Text>
-            </div>
-
-            {questionList?.times.map((time, idx) => (
-              <Flex justify="start" webGap={20} key={`time_${idx}`}>
-                <Text webTypo="Heading4">{time.date}</Text>
-                <ColumnLine />
-                <CheckBox
-                  checked={false}
-                  onClick={onClickCheck}
-                  value={['불가능한 시간', '없음']}
-                  type="column"
-                />
-                <ColumnLine />
-                {time.durations.map((duration, idx) => {
-                  const [start, end] = duration.split('-');
-                  return (
-                    <CheckBox
-                      checked={false}
-                      onClick={onClickCheck}
-                      value={[start, `~ ${end}`]}
-                      type="column"
-                      key={`duration_${idx}`}
-                    />
-                  );
-                })}
-              </Flex>
-            ))}
-          </Flex>
+          <Part
+            register={register}
+            watch={watch}
+            setValue={setValue}
+            handleSubmit={handleSubmit}
+            questionList={questionList}
+          />
+          {/* 면접 날짜 */}
+          <Schedule
+            register={register}
+            watch={watch}
+            setValue={setValue}
+            handleSubmit={handleSubmit}
+            questionList={questionList}
+            onClickCheck={onClickCheck}
+          />
           <Button variant="default">제출하기</Button>
           <Text webTypo="Label3" paletteColor="Gray3" margin="80px 0 56px 0">
             © 2016-2023 Ceos ALL RIGHTS RESERVED.
@@ -429,20 +166,20 @@ const TopMargin = styled.div`
   }
 `;
 
-const RowLine = styled.div`
+export const RowLine = styled.div`
   width: 856px;
   height: 2px;
   background: #e9ebef;
   margin: 24px 0;
 `;
 
-const ColumnLine = styled.div`
+export const ColumnLine = styled.div`
   width: 2px;
   height: 70px;
   background-color: ${theme.palette.Gray2};
 `;
 
-const Section = styled(Flex)`
+export const Section = styled(Flex)`
   flex-direction: column;
   align-items: start;
   width: 856px;
